feat: add environment variable support for database connection
- Replace hardcoded MongoDB URL with process.env.DB_URL
- Add dotenv configuration loading
- Implement error handling for database connection
- Add connection success logging
