+++ Backend/conn/db.js
@@ -1,6 +1,13 @@
 const mongoose = require("mongoose");
+require("dotenv").config();
+
 async function ConnectMongo() {
-    await mongoose.connect('mongodb://127.0.0.1:27017/FastShop');
+    try {
+        await mongoose.connect(process.env.DB_URL);
+        console.log("Connected to MongoDB ");
+    } catch (error) {
+        console.log("MongoDB connection error:", error);
+    }
 }
 
 module.exports = ConnectMongo;
